Changed from SDL_Surface* to Bitmap, Improved some code

+++ source/EncoXD/EncoShared/Render/Shader.d
@@ -2,7 +2,7 @@ module Enco.Shared.Render.Shader;
 
 import EncoShared;
 
-enum ShaderType : int
+enum ShaderType : u8
 {
 	Vertex, TessControl, TessEvaluation, Geometry, Fragment
 }
